:art: refactor: move handler interfaces to their own modules
