+++ api/app/domain/compute_engines/handlers/analytics_otf_handler.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+from app.domain.models.dataset import DatasetQuery
+
+
+class AnalyticsOTFHandler(ABC):
+    @abstractmethod
+    async def handle(self, aoi_type, aoi_ids, query: DatasetQuery):
+        raise NotImplementedError()
